+++ src/App.jsx
@@ -1,31 +1,9 @@
-import { useState } from "react";
-import TodoForm from "./components/TodoForm";
 import TodoList from "./components/TodoList";
+import TodoForm from "./components/TodoForm";
+import { useTodos } from "./hooks/useTodos";
 
 function App() {
-  const [todos, setTodos] = useState([]);
-
-  const addTodo = (value) => {
-    const newTodo = {
-      id: Date.now(),
-      value,
-      completed: false,
-    };
-
-    setTodos([...todos, newTodo]);
-  };
-
-  const deleteTodo = (id) => {
-    setTodos(todos.filter((todo) => todo.id !== id));
-  };
-
-  const toggleTodo = (id) => {
-    setTodos(
-      todos.map((todo) =>
-        todo.id === id ? { ...todo, completed: !todo.completed } : todo,
-      ),
-    );
-  };
+  const { todos, addTodo, deleteTodo, toggleTodo } = useTodos();
 
   return (
     <div>
