+++ src/utils/storage.js
@@ -1,0 +1,8 @@
+export const getStorage = (key) => {
+  const data = localStorage.getItem(key);
+  return data ? JSON.parse(data) : [];
+};
+
+export const setStorage = (key, data) => {
+  localStorage.setItem(key, JSON.stringify(data));
+};
